chore (components) split styles from components
