+++ src/components/FloatingActionButton/index.js
@@ -7,10 +7,7 @@ import {
   Platform,
   TouchableWithoutFeedback,
 } from 'react-native';
-
-const BUTTON_WIDTH         = 72;
-const BUTTON_HEIGHT        = 72;
-const BUTTON_BORDER_RADIUS = 36;
+import styles from './styles';
 
 // Button which will have an absolute position on the bottom right corner
 class FloatingActionButton extends Component {
@@ -34,33 +31,5 @@ class FloatingActionButton extends Component {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    width: BUTTON_WIDTH,
-    height: BUTTON_HEIGHT,
-    borderRadius: BUTTON_BORDER_RADIUS,
-    position: 'absolute',
-    bottom: 10,
-    right: 10,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  buttonView: {
-    width: BUTTON_WIDTH,
-    height: BUTTON_HEIGHT,
-    borderRadius: BUTTON_BORDER_RADIUS,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  floatingActionButtonText: {
-    fontSize: 16,
-  },
-  plusIcon: {
-    width: BUTTON_WIDTH,
-    height: BUTTON_HEIGHT
-  }
-});
-
 
 module.exports = FloatingActionButton
