+++ src/components/NativePicker/styles.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  modalPicker: {
+    borderWidth: 0,
+    alignItems: 'flex-start'
+  },
+});
+
+export default styles
