+++ src/components/Thumbnail/styles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flexDirection: 'column',
+    alignItems: 'flex-start',
+    justifyContent: 'flex-end',
+  },
+  removeImage: {
+    position: 'absolute',
+    top: 0,
+    right: 0,
+    height: 24,
+    width: 24,
+  }
+});
+
+export default styles
